Remove upgrade *.sql
GPDB will require a different set of upgrade scripts than upstream, as
we are coming from a custom version of Orafce to something that much
more closely resembles upstream 3.7.

We still need to think about how to upgrade to the new orafce version in
GPDB, but these scripts are not the solution.